remove duped cypress tests. env vars to indicate if screenshots should be taken or not

+++ client/tests/e2e/specs/ci/CardGameUnit.spec.ts
@@ -15,14 +15,16 @@ beforeEach(() => cy.restoreLocalStorage());
 afterEach(() => cy.saveLocalStorage());
 
 //delete actuals screenshots
-describe("deletes-screenshots", () => {
-  it("deletes any actuals for this test before we enter the page", () => {
-    cy.deleteActuals({
-      headedPath: ACTUALS_CARD_GAME_UNIT_SPEC_PATH,
-      headlessPath: ACTUALS_CARD_GAME_UNIT_SPEC_PATH_HEADLESS
+if (Cypress.env("TAKE_SCREENSHOTS") === "yes") {
+  describe("deletes-screenshots", () => {
+    it("deletes any actuals for this test before we enter the page", () => {
+      cy.deleteActuals({
+        headedPath: ACTUALS_CARD_GAME_UNIT_SPEC_PATH,
+        headlessPath: ACTUALS_CARD_GAME_UNIT_SPEC_PATH_HEADLESS
+      });
     });
   });
-});
+}
 
 // load home page and sign up new user
 describe("visits home page", () => {
@@ -83,6 +85,15 @@ describe("adding a card and checking the flip and translation error or success",
     cy.addCard();
   });
 });
+//screenshot the card itself after it is added and loaded after the loading transition
+if (Cypress.env("TAKE_SCREENSHOTS") === "yes") {
+  describe("screenshot-card", () => {
+    it("finds the first card in the list and screenshots the element", () => {
+      // default theme for new user is light theme
+      cy.get("div.notification.is-light").should("have.length", 1).screenshot();
+    });
+  });
+}
 
 describe("checking if clicking the choice button answer results in correct and incorrect increment will increment", () => {
